+++ constants/index.js
@@ -1,3 +1,11 @@
+export const HIT = {
+    head: 30,
+    body: 25,
+    foot: 20,
+};
+
+export const ATTACK = ['head', 'body', 'foot'];
+
 export const LOGS = {
     start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
     end: [
