--- elements.js
@@ -1,6 +1,0 @@
-const formFight = document.querySelector('.control');
-const arenas = document.querySelector('.arenas');
-const fightButton = formFight.querySelector('.button');
-const chat = document.querySelector('.chat');
-
-export { formFight, arenas, fightButton, chat };
