--- fight.js
@@ -1,32 +1,0 @@
-import { HIT, ATTACK } from './values.js';
-import { getRandom } from './utils.js';
-import { formFight } from './elements.js';
-
-function enemyAttack() {
-    const hit = ATTACK[getRandom(3) - 1];
-    const defence = ATTACK[getRandom(3) - 1];
-
-    return {
-        value: getRandom(HIT[hit]),
-        hit,
-        defence,
-    }
-}
-
-function playerAttack() {
-    const attack = {};
-    for (let item of formFight) {
-        if (item.checked && item.name === 'hit') {
-            attack.value = getRandom(HIT[item.value]);
-            attack.hit = item.value;
-        }
-        if (item.checked && item.name === 'defence') {
-            attack.defence = item.value;
-        }
-        item.checked = false;
-    }
-
-    return attack;
-}
-
-export { enemyAttack, playerAttack };
